Update astyle to 2.05.1, add astyle Linux binaries
Unfortunately, different versions of astyle produce slighty different
formatting, so it's important that everyone uses the same version.

Thus it makes sense to provide astyle binaries for Win32 and
Linux x86 and x86_64 (to prevent usage of outdated versions from package
managers etc)

Most probably it would be easy to add an OSX astyle binary as well and
call that from astyle-code.sh if applicable.
I don't have a Mac, though, so someone else will have to do it ;-)